Update fake data structure
