+++ app/controllers/complaints_controller.rb
@@ -1,5 +1,5 @@
 class ComplaintsController < ApplicationController
   def index
-    @complaints = FakeData::Complaints.generate_response[:data]
+    @complaints = FakeData::ApiResponse.generate_hses_response[:data]
   end
 end
